+++ documents/changeLog.md
@@ -1,6 +1,7 @@
 ## 01 August 2026
 
 - Added viewing options to QR code viewer: Stylised, Standard and HDR (on supported devices)
+- Fixed invite URL and QR code to consistently use roboref.fyi links
 
 ## 31 July 2026